+++ src/s390x/mod.rs
@@ -16,4 +16,6 @@ pub mod dasd;
 pub mod zipl;
 
 pub use dasd::{image_copy_s390x, prepare_dasd};
+pub use eckd::dasd_try_get_sector_size;
 pub use zipl::install_bootloader;
+mod eckd;
